requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pypdf==5.4.0
python-docx==1.1.2
langchain-text-splitters==0.3.8
